chore: rename search params to limits
Also move it to a new module for search called `limits`.

bench: 845875
