+++ engine/src/search/limits.rs
@@ -1,0 +1,76 @@
+use std::{
+    fmt::Display,
+    time::{Duration, Instant},
+};
+
+use chess::board::Board;
+use uci_parser::UciSearchOptions;
+
+use crate::defs::MAX_DEPTH;
+
+/// Input parameters for the search.
+#[derive(Clone, Debug)]
+pub struct SearchLimits {
+    pub max_depth: u8,
+    pub start_time: Instant,
+    pub soft_timeout: Duration,
+    pub hard_timeout: Duration,
+    pub max_nodes: u64,
+}
+
+impl Default for SearchLimits {
+    fn default() -> Self {
+        SearchLimits {
+            max_depth: MAX_DEPTH,
+            start_time: Instant::now(),
+            soft_timeout: Duration::MAX,
+            hard_timeout: Duration::MAX,
+            max_nodes: u64::MAX,
+        }
+    }
+}
+
+impl SearchLimits {
+    /// Creates a new set of search parameters from the UCI options and the current board.
+    pub fn new(uci_options: &UciSearchOptions, board: &Board) -> Self {
+        let mut params = Self::default();
+        if let Some(depth) = uci_options.depth {
+            params.max_depth = depth as u8;
+        }
+
+        if let Some(nodes) = uci_options.nodes {
+            params.max_nodes = nodes as u64;
+        }
+
+        if let Some(time) = uci_options.movetime {
+            params.soft_timeout = time;
+            params.hard_timeout = time;
+        } else {
+            let (time, increment) = if board.side_to_move().is_white() {
+                (uci_options.wtime, uci_options.winc)
+            } else {
+                (uci_options.btime, uci_options.binc)
+            };
+
+            // do we have valid time
+            if let Some(time) = time {
+                // TODO: How can we tune these params?
+                let inc = increment.unwrap_or(Duration::ZERO) / 2;
+                params.soft_timeout = time / 20 + inc;
+                params.hard_timeout = time / 5 + inc;
+            }
+        }
+
+        params
+    }
+}
+
+impl Display for SearchLimits {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "max depth {} start_time {:?} soft_timeout {:?} hard_timeout {:?}",
+            self.max_depth, self.start_time, self.soft_timeout, self.hard_timeout
+        )
+    }
+}
